src/services/django-config/requirements.txt
# Django Requirements for Irrigation Management System
# Install with: pip install -r requirements.txt

# Core Django packages
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
djangorestframework-gis>=1.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
django-environ>=0.10.0  # Environment variable management

# File handling
Pillow>=9.5.0  # Image processing
django-storages>=1.13.0  # File storage backends

# GIS dependencies (system level packages needed)
# Install these via system package manager:
# Ubuntu/Debian: sudo apt-get install gdal-bin libgdal-dev libgeos-dev libproj-dev
# CentOS/RHEL: sudo yum install gdal-devel geos-devel proj-devel
# macOS: brew install gdal geos proj

# Authentication & Security
djangorestframework-simplejwt>=5.2.0  # JWT authentication
django-rest-auth>=0.9.5

# API Documentation (optional)
drf-spectacular>=0.26.0  # OpenAPI schema generation

# Development dependencies
django-debug-toolbar>=4.1.0  # Development debugging
django-extensions>=3.2.0  # Additional Django management commands

# File export dependencies
reportlab>=3.6.0  # PDF generation for BOQ exports
openpyxl>=3.1.0  # Excel file generation
xlsxwriter>=3.1.0  # Alternative Excel writer

# Geospatial file processing
Fiona>=1.9.0  # Shapefile reading
Shapely>=2.0.0  # Geometric operations
geopandas>=0.13.0  # Geospatial data analysis

# Task queue (optional for heavy processing)
celery>=5.3.0  # Asynchronous task processing
redis>=4.5.0  # Celery broker

# Production dependencies
gunicorn>=20.1.0  # WSGI HTTP Server
whitenoise>=6.5.0  # Static file serving